Design correction
Corrected the design of the pawn vehicle. This is different from the other because it is not a very vehicle, so the pawn is an extension of a citizen but with the capability of a vehicle. It can transport a citizen from an HOME address to a WORK address and viceversa.